Created TroiApiWrapper.js
